# Performance improvements for VectorDict