Implements a fast track for dot products when only the intersection of the keys is needed.
Also a bug fix for cases with mask.
This addresses concerns of [low performance](https://github.com/creme-ml/creme/pull/316#issuecomment-684126028) shared by @gbolmier, below is updated benchmarks in this branch.

```py
from creme import utils

x = {i: i for i in range(100)}
xx = {i: i for i in range(1_000)}
y = utils.VectorDict(x)
yy = utils.VectorDict(xx)

%timeit utils.math.dot(x, x)
# 12.2 µs ± 120 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

%timeit utils.math.dot(xx, x)
# 12.1 µs ± 95.6 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

%timeit y @ y
# 8.93 µs ± 126 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

%timeit yy @ y
# 8.93 µs ± 102 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```