@gbolmier 2020-09-15
Thx for your work! Do you think there is still room for improvement? Your original implementation reported ≈ 10x improvement for 100 features (https://github.com/creme-ml/creme/pull/315#issuecomment-656549424). Or does `VectorDict` genericity prevent better performance?