@MaxHalford 2020-09-15
I'll merge this because it's an improvement. You can keep discussing here or elsewhere :)