@MaxHalford 2020-09-16
@brcharron I believe some of the unit tests in `test_vectordict.py` are not working, can you check?