@brcharron 2020-09-16
@MaxHalford Thanks for merging, the tests pass on my side, did you rebuild Cython maybe?
Which assert broke? I added some to catch a bug that was in `master`.